+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 
 
 
+## [2.4.0] - 2022-04-08
+[Full Changelog](https://github.com/TechNobre/PowerUtils.GuardClauses.Validations/compare/v2.3.0...v2.4.0)
+
+
+### Enhancements
+- Removed `PowerUtils.Text` dependency;
+
+
+### New Features
+- Added Guard `Guard.Validate.IfOutOfRange()` for dateTime;
+- Added Guard `Guard.Validate.IfOutOfRange()` for numbers;
+
+
+
+
 ## [2.3.0] - 2022-04-05
 [Full Changelog](https://github.com/TechNobre/PowerUtils.GuardClauses.Validations/compare/v2.2.1...v2.3.0)